requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.12
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
numpy==2.2.4
openai==1.69.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
protobuf==6.30.2
pydantic==2.11.1
pydantic_core==2.33.0
pymilvus==2.5.6
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1